requirements.txt
Django==4.2.7
django_restql==0.15.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
Requests==2.31.0
user_agents==2.2.0
mysqlclient==2.2.0
django-cors-headers
websockets
